add authguard pages
