+++ app/info/page.tsx
@@ -1,3 +1,4 @@
+import { AuthGuard } from "@/components/AuthGuard";
 import { List } from "@/components/List";
 import { getAllAnnouncementClasses } from "@/db/getAllAnnouncementClasses";
 import { getAllAnnouncements } from "@/db/getAllAnnouncements";
@@ -19,7 +20,7 @@ export default async function Info() {
     return { id: i, param: id, date, title, subtext: formattedClasses };
   });
   return (
-    <>
+    <AuthGuard filter={{ canManage: true }}>
       <h1 className={shared.title}>お知らせ一覧</h1>
       <List
         items={items}
@@ -27,6 +28,6 @@ export default async function Info() {
         link="/info/"
         query="?from=/info"
       />
-    </>
+    </AuthGuard>
   );
 }
